Crate rusoto_kafka

The operations for managing an Amazon MSK cluster.

If you're using the service, you're probably looking for KafkaClient and Kafka.

Structs

BrokerNodeGroupInfo
Describes the setup to be used for Kafka broker nodes in the cluster.

BrokerNodeInfo
BrokerNodeInfo

BrokerSoftwareInfo
Information about the current software installed on the cluster.

ClusterInfo
Returns information about a cluster.

CreateClusterRequest

CreateClusterResponse

DeleteClusterRequest

DeleteClusterResponse

DescribeClusterRequest

DescribeClusterResponse

EBSStorageInfo
Contains information about the EBS storage volumes attached to Kafka broker nodes.

EncryptionAtRest
The data volume encryption details.

EncryptionInfo
Includes encryption-related information, such as the AWS KMS key used for encrypting data at rest.

GetBootstrapBrokersRequest

GetBootstrapBrokersResponse

KafkaClient
A client for the Kafka API.

KafkaError
Returns information about an error.

ListClustersRequest

ListClustersResponse

ListNodesRequest

ListNodesResponse

NodeInfo
The node information object.

StorageInfo
Contains information about storage volumes attached to MSK broker nodes.

ZookeeperNodeInfo
Zookeeper node information.

Enums

CreateClusterError
Errors returned by CreateCluster

DeleteClusterError
Errors returned by DeleteCluster

DescribeClusterError
Errors returned by DescribeCluster

GetBootstrapBrokersError
Errors returned by GetBootstrapBrokers

ListClustersError
Errors returned by ListClusters

ListNodesError
Errors returned by ListNodes

Traits

Kafka
Trait representing the capabilities of the Kafka API. Kafka clients implement this trait.