Struct rusoto_kafka::KafkaClient

pub struct KafkaClient { /* fields omitted */ }

A client for the Kafka API.

Methods

impl KafkaClient

pub fn new(region: Region) -> KafkaClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> KafkaClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,