Trait rusoto_kafka::Kafka

pub trait Kafka {
    fn create_cluster(
        &self, 
        input: CreateClusterRequest
    ) -> RusotoFuture<CreateClusterResponse, CreateClusterError>;

    fn delete_cluster(
        &self, 
        input: DeleteClusterRequest
    ) -> RusotoFuture<DeleteClusterResponse, DeleteClusterError>;

    fn describe_cluster(
        &self, 
        input: DescribeClusterRequest
    ) -> RusotoFuture<DescribeClusterResponse, DescribeClusterError>;

    fn get_bootstrap_brokers(
        &self, 
        input: GetBootstrapBrokersRequest
    ) -> RusotoFuture<GetBootstrapBrokersResponse, GetBootstrapBrokersError>;

    fn list_clusters(
        &self, 
        input: ListClustersRequest
    ) -> RusotoFuture<ListClustersResponse, ListClustersError>;

    fn list_nodes(
        &self, 
        input: ListNodesRequest
    ) -> RusotoFuture<ListNodesResponse, ListNodesError>;
}

Trait representing the capabilities of the Kafka API. Kafka clients implement this trait.

Required methods

fn create_cluster(
&self,
input: CreateClusterRequest
) -> RusotoFuture<CreateClusterResponse, CreateClusterError>

Creates a new MSK cluster.

fn delete_cluster(
&self,
input: DeleteClusterRequest
) -> RusotoFuture<DeleteClusterResponse, DeleteClusterError>

Deletes the MSK cluster specified by the Amazon Resource Name (ARN) in the request.

fn describe_cluster(
&self,
input: DescribeClusterRequest
) -> RusotoFuture<DescribeClusterResponse, DescribeClusterError>

Returns a description of the MSK cluster whose Amazon Resource Name (ARN) is specified in the request.

fn get_bootstrap_brokers(
&self,
input: GetBootstrapBrokersRequest
) -> RusotoFuture<GetBootstrapBrokersResponse, GetBootstrapBrokersError>

A list of brokers that a client application can use to bootstrap.

fn list_clusters(
&self,
input: ListClustersRequest
) -> RusotoFuture<ListClustersResponse, ListClustersError>

Returns a list of clusters in an account.

fn list_nodes(
&self,
input: ListNodesRequest
) -> RusotoFuture<ListNodesResponse, ListNodesError>

Returns a list of the broker nodes in the cluster.

Loading content...