Crate rusoto_kafka

           <p>The operations for managing an Amazon MSK cluster.</p>

If you're using the service, you're probably looking for KafkaClient and Kafka.

Structs

BrokerEBSVolumeInfo
        <p>Specifies the EBS volume upgrade information. The broker identifier must be set to the keyword ALL. This means the changes apply to all the brokers in the cluster.</p>

BrokerNodeGroupInfo
        <p>Describes the setup to be used for Kafka broker nodes in the cluster.</p>

BrokerNodeInfo
        <p>BrokerNodeInfo</p>

BrokerSoftwareInfo
        <p>Information about the current software installed on the cluster.</p>

ClientAuthentication
        <p>Includes all client authentication information.</p>

ClusterInfo
        <p>Returns information about a cluster.</p>

ClusterOperationInfo
        <p>Returns information about a cluster operation.</p>

Configuration
        <p>Represents an MSK Configuration.</p>

ConfigurationInfo
        <p>Specifies the configuration to use for the brokers.</p>

ConfigurationRevision
        <p>Describes a configuration revision.</p>

CreateClusterRequest

CreateClusterResponse

CreateConfigurationRequest

CreateConfigurationResponse

DeleteClusterRequest

DeleteClusterResponse

DescribeClusterOperationRequest

DescribeClusterOperationResponse

DescribeClusterRequest

DescribeClusterResponse

DescribeConfigurationRequest

DescribeConfigurationResponse

DescribeConfigurationRevisionRequest

DescribeConfigurationRevisionResponse

EBSStorageInfo
        <p>Contains information about the EBS storage volumes attached to Kafka broker nodes.</p>

EncryptionAtRest
        <p>The data-volume encryption details.</p>

EncryptionInTransit
        <p>The settings for encrypting data in transit.</p>

EncryptionInfo
        <p>Includes encryption-related information, such as the AWS KMS key used for encrypting data at rest and whether you want MSK to encrypt your data in transit.</p>

ErrorInfo
        <p>Returns information about an error state of the cluster.</p>

GetBootstrapBrokersRequest

GetBootstrapBrokersResponse

KafkaClient
A client for the Kafka API.

ListClusterOperationsRequest

ListClusterOperationsResponse

ListClustersRequest

ListClustersResponse

ListConfigurationRevisionsRequest

ListConfigurationRevisionsResponse

ListConfigurationsRequest

ListConfigurationsResponse

ListNodesRequest

ListNodesResponse

ListTagsForResourceRequest

ListTagsForResourceResponse

MutableClusterInfo
        <p>Information about cluster attributes that can be updated via update APIs.</p>

NodeInfo
        <p>The node information object.</p>

StorageInfo
        <p>Contains information about storage volumes attached to MSK broker nodes.</p>

TagResourceRequest

Tls
        <p>Details for client authentication using TLS.</p>

UntagResourceRequest

UpdateBrokerStorageRequest

UpdateBrokerStorageResponse

UpdateClusterConfigurationRequest

UpdateClusterConfigurationResponse

ZookeeperNodeInfo
        <p>Zookeeper node information.</p>

Enums

CreateClusterError
Errors returned by CreateCluster

CreateConfigurationError
Errors returned by CreateConfiguration

DeleteClusterError
Errors returned by DeleteCluster

DescribeClusterError
Errors returned by DescribeCluster

DescribeClusterOperationError
Errors returned by DescribeClusterOperation

DescribeConfigurationError
Errors returned by DescribeConfiguration

DescribeConfigurationRevisionError
Errors returned by DescribeConfigurationRevision

GetBootstrapBrokersError
Errors returned by GetBootstrapBrokers

ListClusterOperationsError
Errors returned by ListClusterOperations

ListClustersError
Errors returned by ListClusters

ListConfigurationRevisionsError
Errors returned by ListConfigurationRevisions

ListConfigurationsError
Errors returned by ListConfigurations

ListNodesError
Errors returned by ListNodes

ListTagsForResourceError
Errors returned by ListTagsForResource

TagResourceError
Errors returned by TagResource

UntagResourceError
Errors returned by UntagResource

UpdateBrokerStorageError
Errors returned by UpdateBrokerStorage

UpdateClusterConfigurationError
Errors returned by UpdateClusterConfiguration

Traits

Kafka
Trait representing the capabilities of the Kafka API. Kafka clients implement this trait.