Trait rusoto_kafka::Kafka

pub trait Kafka {
    fn create_cluster(
        &self, 
        input: CreateClusterRequest
    ) -> RusotoFuture<CreateClusterResponse, CreateClusterError>;

    fn create_configuration(
        &self, 
        input: CreateConfigurationRequest
    ) -> RusotoFuture<CreateConfigurationResponse, CreateConfigurationError>;

    fn delete_cluster(
        &self, 
        input: DeleteClusterRequest
    ) -> RusotoFuture<DeleteClusterResponse, DeleteClusterError>;

    fn describe_cluster(
        &self, 
        input: DescribeClusterRequest
    ) -> RusotoFuture<DescribeClusterResponse, DescribeClusterError>;

    fn describe_cluster_operation(
        &self, 
        input: DescribeClusterOperationRequest
    ) -> RusotoFuture<DescribeClusterOperationResponse, DescribeClusterOperationError>;

    fn describe_configuration(
        &self, 
        input: DescribeConfigurationRequest
    ) -> RusotoFuture<DescribeConfigurationResponse, DescribeConfigurationError>;

    fn describe_configuration_revision(
        &self, 
        input: DescribeConfigurationRevisionRequest
    ) -> RusotoFuture<DescribeConfigurationRevisionResponse, DescribeConfigurationRevisionError>;

    fn get_bootstrap_brokers(
        &self, 
        input: GetBootstrapBrokersRequest
    ) -> RusotoFuture<GetBootstrapBrokersResponse, GetBootstrapBrokersError>;

    fn list_cluster_operations(
        &self, 
        input: ListClusterOperationsRequest
    ) -> RusotoFuture<ListClusterOperationsResponse, ListClusterOperationsError>;

    fn list_clusters(
        &self, 
        input: ListClustersRequest
    ) -> RusotoFuture<ListClustersResponse, ListClustersError>;

    fn list_configuration_revisions(
        &self, 
        input: ListConfigurationRevisionsRequest
    ) -> RusotoFuture<ListConfigurationRevisionsResponse, ListConfigurationRevisionsError>;

    fn list_configurations(
        &self, 
        input: ListConfigurationsRequest
    ) -> RusotoFuture<ListConfigurationsResponse, ListConfigurationsError>;

    fn list_nodes(
        &self, 
        input: ListNodesRequest
    ) -> RusotoFuture<ListNodesResponse, ListNodesError>;

    fn list_tags_for_resource(
        &self, 
        input: ListTagsForResourceRequest
    ) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>;

    fn tag_resource(
        &self, 
        input: TagResourceRequest
    ) -> RusotoFuture<(), TagResourceError>;

    fn untag_resource(
        &self, 
        input: UntagResourceRequest
    ) -> RusotoFuture<(), UntagResourceError>;

    fn update_broker_storage(
        &self, 
        input: UpdateBrokerStorageRequest
    ) -> RusotoFuture<UpdateBrokerStorageResponse, UpdateBrokerStorageError>;

    fn update_cluster_configuration(
        &self, 
        input: UpdateClusterConfigurationRequest
    ) -> RusotoFuture<UpdateClusterConfigurationResponse, UpdateClusterConfigurationError>;
}

Trait representing the capabilities of the Kafka API. Kafka clients implement this trait.

Required methods

fn create_cluster(
&self,
input: CreateClusterRequest
) -> RusotoFuture<CreateClusterResponse, CreateClusterError>

        <p>Creates a new MSK cluster.</p>

fn create_configuration(
&self,
input: CreateConfigurationRequest
) -> RusotoFuture<CreateConfigurationResponse, CreateConfigurationError>

        <p>Creates a new MSK configuration.</p>

fn delete_cluster(
&self,
input: DeleteClusterRequest
) -> RusotoFuture<DeleteClusterResponse, DeleteClusterError>

        <p>Deletes the MSK cluster specified by the Amazon Resource Name (ARN) in the request.</p>

fn describe_cluster(
&self,
input: DescribeClusterRequest
) -> RusotoFuture<DescribeClusterResponse, DescribeClusterError>

        <p>Returns a description of the MSK cluster whose Amazon Resource Name (ARN) is specified in the request.</p>

fn describe_cluster_operation(
&self,
input: DescribeClusterOperationRequest
) -> RusotoFuture<DescribeClusterOperationResponse, DescribeClusterOperationError>

        <p>Returns a description of the cluster operation specified by the ARN.</p>

fn describe_configuration(
&self,
input: DescribeConfigurationRequest
) -> RusotoFuture<DescribeConfigurationResponse, DescribeConfigurationError>

        <p>Returns a description of this MSK configuration.</p>

fn describe_configuration_revision(
&self,
input: DescribeConfigurationRevisionRequest
) -> RusotoFuture<DescribeConfigurationRevisionResponse, DescribeConfigurationRevisionError>

        <p>Returns a description of this revision of the configuration.</p>

fn get_bootstrap_brokers(
&self,
input: GetBootstrapBrokersRequest
) -> RusotoFuture<GetBootstrapBrokersResponse, GetBootstrapBrokersError>

        <p>A list of brokers that a client application can use to bootstrap.</p>

fn list_cluster_operations(
&self,
input: ListClusterOperationsRequest
) -> RusotoFuture<ListClusterOperationsResponse, ListClusterOperationsError>

        <p>Returns a list of all the operations that have been performed on the specified MSK cluster.</p>

fn list_clusters(
&self,
input: ListClustersRequest
) -> RusotoFuture<ListClustersResponse, ListClustersError>

        <p>Returns a list of all the MSK clusters in the current Region.</p>

fn list_configuration_revisions(
&self,
input: ListConfigurationRevisionsRequest
) -> RusotoFuture<ListConfigurationRevisionsResponse, ListConfigurationRevisionsError>

        <p>Returns a list of all the MSK configurations in this Region.</p>

fn list_configurations(
&self,
input: ListConfigurationsRequest
) -> RusotoFuture<ListConfigurationsResponse, ListConfigurationsError>

        <p>Returns a list of all the MSK configurations in this Region.</p>

fn list_nodes(
&self,
input: ListNodesRequest
) -> RusotoFuture<ListNodesResponse, ListNodesError>

        <p>Returns a list of the broker nodes in the cluster.</p>

fn list_tags_for_resource(
&self,
input: ListTagsForResourceRequest
) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>

        <p>Returns a list of the tags associated with the specified resource.</p>

fn tag_resource(
&self,
input: TagResourceRequest
) -> RusotoFuture<(), TagResourceError>

        <p>Adds tags to the specified MSK resource.</p>

fn untag_resource(
&self,
input: UntagResourceRequest
) -> RusotoFuture<(), UntagResourceError>

        <p>Removes the tags associated with the keys that are provided in the query.</p>

fn update_broker_storage(
&self,
input: UpdateBrokerStorageRequest
) -> RusotoFuture<UpdateBrokerStorageResponse, UpdateBrokerStorageError>

        <p>Updates the EBS storage associated with MSK brokers.</p>

fn update_cluster_configuration(
&self,
input: UpdateClusterConfigurationRequest
) -> RusotoFuture<UpdateClusterConfigurationResponse, UpdateClusterConfigurationError>

        <p>Updates the cluster with the configuration that is specified in the request body.</p>

Loading content...